Ensure the "Apply" button on needles table is always clickable
Before this change the "Apply" button next to the date/time picker could
overflow at certain viewport sizes and was thus not clickable anymore. This
change solves this by giving those columns more space. It also fixes the
alignment of the buttons on the right column when columns are collapsed.

I noticed this problem when working on
https://progress.opensuse.org/issues/167611 but this change will not fix
that test.