Add loading indicator to the test liveview
This PR adds a loading animation indicating that the liveview is being loaded. Currently only a empty box is displayed, visible in the webui before liveview content is received by the responsible canvas element.
Test cases are also added to confirm the elements are displayed and hidden as expected.

Related Issue: https://progress.opensuse.org/issues/134840