Avoid scheduling jobs if not all parallel jobs are ready
Some jobs of a parallel cluster can be blocked (by a chained parent or by a
pending Gru task) while some jobs can be scheduled. Before this change the
scheduler assigns the jobs that can be scheduled which creates a half-
scheduled parallel cluster.

This is particularly problematic when `PARALLEL_ONE_HOST_ONLY=1` and
`git_auto_update = yes` are used because then repairing half-scheduled
clusters is more challenging and the likeliness that a cluster is partially
blocked is higher. In theory this is also problematic when jobs within the
parallel cluster depend on different asset downloads.

With this change the scheduler skips the whole cluster until all jobs can
be scheduled to avoid half-scheduled clusters. Judging by the previous code
and its comments this was already intended but the code didn't actually
work correctly.

I tested the rewritten version of the code by creating a half-blocked
cluster manually and observed the behavior of the scheduler before and
after this commit.

Related ticket: https://progress.opensuse.org/issues/169342