Make 'Broken' workers show as 'Unavailable' instead
This commit displays "Broken" workers as instead "Unavailable",
as the previous wording was e.g. not including workers which where
unavailable due to 'too high load'.
For the same reason the 'Error' title of the worker states hover popup is replaced with
'Details' and the 'too high load' output is enhanced.

Related Ticket: https://progress.opensuse.org/issues/164084