Check existing job groups in openqa-load-templates
The `openqa-load-templates` script now checks if a job groups is already present when importing new template data. If a group already exists, there will be no attempt of creating it, but associated templates will still be loaded if the `--update` option is provided.
If the `--update` option is not provided, and a duplicate job group is found, the import will fail.

Related Issue: https://progress.opensuse.org/issues/174808