Update authentication headers when following redirection via user agent
The openQA user agent code so far preserves manually set headers including
the `X-API-…` headers used for authentication. This is probably not
required and has the problematic side-effect that those headers are not
updated when following redirections. That means authentication fails when a
redirection is in place. This is the case when querying assets of a job as
done by the cache service.

In order to make the cache service work when authentication for assets is
enabled via 4c8ed396839e679b5e364c3a625f056204cf424c this change updates
the `X-API-…` headers regardless of whether they are already present. Any
custom values will be overridden but preserving custom values is not
required anyways. (The accepted header will still be preserved.)

Related ticket: https://progress.opensuse.org/issues/174154