Make key/secret auth work in `/api/v1/auth` route via `auth_request`
When the `/api/v1/auth` route is used via the NGINX configuration
`auth_request` the request URL is not the URL relevant for authentication
because the request URL/path will always be `/api/v1/auth`. Instead the
`X-Original-URI` is used to specify the actual path as done in the NGINX
config template.

Related ticket: https://progress.opensuse.org/issues/170380