Avoid nested transaction when creating jobs from settings
The transaction in the function `create_from_settings` was introduced by
b03aa8d6c6. At this point this function was likely not always called within
`txn_do`. It is now, though. So I suppose this extra nested transaction is
not useful anymore.

This nested transaction also doesn't act as a useful save point as the
errors we saw in https://progress.opensuse.org/issues/177048 indicate.