Test saving new needle using image from existing needle
The needle editor test so far only tested saving new needles based on the
screenshot. This new test also tests whether it works when basing the image
of the new needle on an existing one and checks that by comparing the image
checksums.

Related ticket: https://progress.opensuse.org/issues/177564