Use Feature::Compat::Try in JobTemplate controller
This one is a bit more complicated because it uses several `return` statements,
but just for an early return of the control flow. With Try::Tiny it would just
leave the `try` block.
In that case we can use an inner extra block and use `last` to leave the
`try` without returning from the sub.

Issue: https://progress.opensuse.org/issues/176862