openqa-dump-templates: Reproducible output
The output file in JSON format is now reproducible, because all keys
will be sorted before the JSON file is written.
The JSON spec states that objects are unordered, but for the purpose of
storing the JSON version of the dump in a VCS it helps that the
difference between consecutive commits is minimal.