Adapt use of `local-npm-registry` to version 1.1.0
As of version 1.1.0 `local-npm-registry` always invokes `npm` with
`--no-package-lock`. This means we no longer need to clean up
`package-lock.json` and also must not do it because the file will otherwise
be missing later on. This change adds the `--no-package-lock` explicitly so
we get the same behavior regardless of the `local-npm-registry` version.

Related ticket: https://progress.opensuse.org/issues/177901