Fix non-custom OAuth2 providers
The fb002508a commit introduced usage of 'id_from' config parameter. And
while the commit messsage says it defaults to the old value 'id', in
fact the parameter is set only for the custom provider, and in all the
other cases it isn't included in $provider_config at all.

Fix this by having default/fallback value also where the variable is
actually used, thus avoiding the need to set it in every supported
provider.

This un-breaks authentication via Github, among other things.

Fixes: fb002508a "Support OpenID Connect better in the OAuth2 'custom' provider"