Simplify WebSockets::Worker::Controller
If $worker is undef, the rest of the statements in the try block
don't get executed and the final return value will be false anyway.

$current_job_id will be only set if there is a $current_job, and all
statements after that will only be true if $current_job_id is defined,
expect some log_trace statements, which we can probably consider not
important.

We can move most of the statements out of the try block because they are
not changing anything in the database or can fail in an expected way.
The "skip any further actions" if statement is only true when $job_id
is defined, and the final statement is only true if $job_id is not
defined, so it can return from the sub directly.

The only code left in the try is reschedule_assigned_jobs().