Link a job only with existing tracking issues
Assume that the referer has a issue_id otherwise do not link the URL in the
comments. Once a ticket has been created the referer should contain an ID
despite where it comes from.
It could go one step further to send a request and check the response but it
doesn't seem to be necessary.

https://progress.opensuse.org/issues/176826