Improve "from" time label in /tests/overview
The previous wording for example "from 2 days ago" could be understood
as showing jobs starting 2 days ago until now. This commit changes the
wording to make it clear that we mean jobs until that specified
timestamp. So in the above example the text would say "at the time of 2
days ago".