dist: Add headers expected by OBS checks
https://build.opensuse.org/requests/1256193 was declined with 'does not
appear to have a license. The file needs to contain a free software
license'. This commit takes over mostly what "osc service runall
format_spec_file" adds automatically with exception of the "(c) 2025"
which is not strictly needed and I would like to avoid. If it turns out
to be necessary after all then we should add it later on.