oauth2: use openQA base_url not url_from to create redirect URI
I'm setting up oauth2 auth for Fedora ATM (up till now we used
OpenID auth, but we're shutting that off). One problem I hit was
that this redirect URI generation used the wrong hostname. Our
staging instance is at https://openqa.stg.fedoraproject.org.
The machine that serves it has the hostname
openqa-lab01.iad2.fedoraproject.org. The redirect_uri was using
the machine's hostname, so it was coming out as
https://openqa-lab01.iad2.fedoraproject.org/login , which is not
going to work.

I can't find any way to tell Mojolicious what host to use for
this kind of `url_for` request, weirdly. It seems like a thing
that should exist, but I can't find it. The next best option I
could find was to use the `base_url` concept that exists only
in openQA (Mojolicious doesn't know about it).

Signed-off-by: Adam Williamson <awilliam@redhat.com>