oauth2: use openQA base_url to create redirect URI if set
I'm setting up oauth2 auth for Fedora ATM (up till now we used
OpenID auth, but we're shutting that off). One problem I hit was
that this redirect URI generation used the wrong hostname. Our
staging instance is at https://openqa.stg.fedoraproject.org.
The machine that serves it has the hostname
openqa-lab01.iad2.fedoraproject.org. The redirect_uri was using
the machine's hostname, so it was coming out as
https://openqa-lab01.iad2.fedoraproject.org/login , which is not
going to work.

This tweaks the redirect URI generation to use the hostname from
the openQA config base_url, if it's set. If not, it falls back
to the hostname from the request.

Signed-off-by: Adam Williamson <awilliam@redhat.com>