Do not show empty groups when has no builds on parent_group_overview
The problem with parent_group_overview is that offers two ways of grouping a
set of data. The design of the query seems to serve the purpose of the
grouped_by_build interface. When it comes to the grouped_by_group, though,
there should be another query to achieve the expected output in the view. That
would likely be, the latest build of each group within the build_limits.

The solution presented in this commit is to visualuze only the groups which
carry some builds from the query, but let user know that more builds are
provided in that parent groups.
To address this problem, another solution would be to add another endpoint in
the API and invoke the resultset from this query.
The idea of modifying the message left out because it would not provide a
better user experience at the end, leaving still a unsatisfied feeling.

The reason for the `clearFlash` is that the panel remains between the change
from one group to another, and otherwise would need to be closed manually.

https://progress.opensuse.org/issues/179296