Adjust limit_builds in parent_group_overview results
With this commit, the limit_builds check is moved into the responsible
template. This gives some flexibility over the resultset but it lets the
compute_build_results to parse, prepare and return all of them. By default the
function query the database for the first 400 jobs unless a bigger value is
requested.

What change is
- *before*: when limit_builds is 10, by_builds lists the 10 latest builds. the
by_group lists the same %result filtering by group. That was ignoring the
latest jobs of the children group.
- *now": when limit_builds is 10, by_builds lists the 10 latest builds but the
by_group lists the latest jobs within the limit_builds which have results for
all the groups.

relative ticket: https://progress.opensuse.org/issues/179296