Add signal guard for schedule_iso minion tasks
Similar to cd99074 to prevent `Job terminated unexpectedly` on results. This
ensures that the minion job will retry in case of interruptions, like network
glitch in the system, without kill the process.

Use the minion guard to retry iso scheduling in case of SIGTERM after the tnx
operation.
Due to the fact that the _schedule_iso_ is used from Iso API, and despite that
likely the guard is useful for those jobs as well, it left out of this
commit. Instead when schedule_iso is called by Iso API the guard is ignored.

https://progress.opensuse.org/issues/108983