Avoid git clone/update operations unless 'scm' is set to 'git'
The introduction of git_auto_update and git_auto_clone seem to
have ignored the existing 'scm' config setting, which is supposed
to control whether openQA does anything git-related to the distri
at all. If 'scm' is not set to 'git', it should not. This was
established when we added the feature of automatically committing
needles created in the editor; that only happens if 'scm' is set
to 'git'. The new git_auto_update and git_auto_clone settings were
added to the 'scm git' section of the config, which is clearly
implied to be context-specific settings that are relevant only
when 'scm' is set to 'git'. But the automatic cloning/update code
never bothers to check if 'scm' is set to 'git', it only checks
the git_auto_* settings.

This is especially problematic because git_auto_* default to yes,
but scm does *not* default to 'git'. Any existing instance which
does not specifically specify 'scm = git' should be assumed not
to want openQA to attempt git operations on its distri, but now
openQA defaults to doing that.

This should solve the problem by consistently checking whether
scm is set to 'git' before doing any git clone operations. We
*could* only have the lowest-level check in _git_clone_all, but
I decided to also add higher-level checks before we reach that
far, to avoid pointless additional work in the case that git is
disabled.

It would be nice to use OpenQA::Git `enabled` for this, but it's
awkward because we often don't have an instance when we need to
check this, and creating one just to do the config check seems
excessive (plus, what path do you use?) We could possibly have a
utility function that is passed a config or an app and does the
`$app->config->{global}->{scm} || '') eq 'git'` check, and have
all these places plus the OpenQA::Git `enabled` sub use that, but
it kinda seems so trivial as to be not worth it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>