Add git_auto_commit setting to control commit of edited needles
Before git_auto_clone and git_auto_update features appeared, we
had the feature where needles created or edited in the webUI
needle editor can be automatically committed back to the distri
repo (assuming it is one). This was controlled only by the
`scm` key in `[global]` being set to `git`, with some details of
the behavior configured in the `[scm git]` section; there was no
specific toggle for the auto-commit behaviour. But now we're also
having git_auto_clone and git_auto_update depend on `scm = git`,
we should make it possible to control the auto_commit behaviour
independently of those two as well. So add a `git_auto_commit`
key to `[scm git]` which controls this. Now things are
consistent: `scm = git` must be set for any of these behaviors
to kick in, and each can be controlled individually with a key
in `[scm git]` (which all default to 'yes').

Also update the docs to explain all this.

Signed-off-by: Adam Williamson <awilliam@redhat.com>