Add linear backoff in hook script
Use the _delay_ which comes from FinalizeResults to override the default Minion
delay behavior and force a linear backoff calculation based on the retries in
case of job failure and until it reaches the max retries. This should offer a
less agressive pattern, giving the Minion tasks more time to recover from
various issues and prevent overloading. Linear backoff is selected over
exponential as the later is already applied to the Minion tasks in general.

https://progress.opensuse.org/issues/181427

Co-authored-by: perlpunk  <tina.mueller@suse.com>