Makefile: Create config folder template only where to be used by users
Only etc/openqa is intended to be used by users. If distributors want
to put things into usr/etc then the according directories should only be
created in the according places but not during generic installation to
prevent problems like distribution specific requirements for such
folders.

This fixes a problem with "ostree" as reported by AdamWill.

Related to
https://progress.opensuse.org/issues/179359#note-33