Ensure only git clone+update features are enabled by default
We still need to keep git features enabled by default.
https://github.com/os-autoinst/openQA/pull/6414 introduced a regression
when only enabling git features when the global key "scm" is set to git
which the comment in openqa.ini states but was actually not true. This
commit changes the default value as well as documentation to correspond
to the comment in openqa.ini as well as what we effectively had
introduced by enabling automatic git clone and update by default while
disabling git_auto_commit to be consistent with previous behaviour. It
is still possible to entirely disable any git features as documented.

Related progress issue: https://progress.opensuse.org/issues/181772