load-templates: job groups: simplify, don't error on group exists
The job group creation API gives a 500 error if the group already
exists. Instead of having another check here, let's attempt to
create the group, then act according to the API response.

Also, as for the other data types, ignore the group and move on
if the group exists and --update wasn't passed, instead of
erroring.

Finally, even if the group exists, continue with the template
string load if --clean was passed. --clean always means to start
fresh and load all data from the template, so we should go ahead
and post the template even if the group already exists. This is
common in --clean mode because we can't fully delete the job
groups, they have properties that aren't stored in the templates.

--clean with JobTemplates doesn't yet work properly
even with this fix, but I'm proposing a full fix for it in
https://progress.opensuse.org/issues/182063 .

Signed-off-by: Adam Williamson <awilliam@redhat.com>