t: load-templates: check harder for what gets loaded
It's not safe to only check one data type in the output. Let's
check all the data types. Making this change exposes an issue: the
machines in the test templates data also exist in the fixtures, so
we skip loading them unless --update or --clean are passed. Let's
give them non-conflicting names to ensure we fully test the data.

Signed-off-by: Adam Williamson <awilliam@redhat.com>