dump-templates: dump job groups as they exist, fix group checks
dump-templates would almost never dump any JobTemplates. First,
we fix two bugs that caused that:

1. `keys %{$options{group}}` triggers perl autovivification and
defines `$options{group}` as an empty array ref even if it was
previously undefined, so later boolean checks for it passed
unexpectedly
2. in 9d311d7, one of the filters in `handle_row` lost a check
for `$options{product}`, so it was incorrectly kicking us out
early unless `--full` or `--product` were passed

With that fixed, though, we default to dumping all job group
and template info both as JobTemplates (the format from before
YAML templates were introduced in
https://progress.opensuse.org/issues/44360) and JobGroups (the format that
can store YAML templates). This is redundant and it's unclear
what we intend load-templates to do with such data.

Instead we'll default to dumping the data in the form it currently
exists on the server. Legacy groups will be dumped to JobTemplates;
"modern" groups will be dumped to JobGroups. This is in line with
the intent of being able to dump the current configuration and
reload it exactly as-is.

We refuse to dump templates with settings or a description as
JobTemplates, as we could not reload them - the legacy API we use for
loading JobTemplates doesn't handle them. This state should be
unreachable without database manipulation, but let's be safe. The
test fixtures *do* start in this 'impossible' state, so we have to
tweak around that, and add a new explicit test that loading YAML
templates with settings/description works. Also test clean and update
with YAML templates: since the recent tweak to load-templates it
works almost correctly (but does not clear existing template
strings for groups not in the data being loaded). We remove
40-templates.json as it has not been used since d601c36.

We also add --convert for if you really want to convert all existing
legacy groups to modern YAML style. It will dump all groups to
JobGroups, and not dump JobTemplates at all.

We also custom sort the job templates before dumping them, to ensure
consistent output, because JSON::XS `canonical` doesn't manage this.

Finally we tweak a test check to use eq_or_diff_text instead of
is_deeply, because the output was much easier to read when debugging
this PR. It's from Test::Differences, which is in Test::Most.

Signed-off-by: Adam Williamson <awilliam@redhat.com>