Add full regex support to job settings filtering
This allows filtering like
`/tests?job_setting=_TEST_ISSUES=(^|,)(33135)(,|$)&…` to really only match
for a specific incident and not accidentally matching incident numbers that
contain the specified incident number.

It is perhaps not ideal to pass the regex unsanitized to to the
database.

Related ticket: https://progress.opensuse.org/issues/156553