Add API route to show all not-ok jobs based on settings
The new API works exactly as its precessor /job_settings/jobs which filters
job_id(s) based on the value of a key or list of keys. In the same logic to
further filter the failing jobs, we can pass use an additional parameter to
query the job's results. OpenQA provides a list of NOT_OK_RESULTS which is
used here to pass to that new parameter.

A reason to have a dedicated function for the new API, as opposed to reuse
the existing is that the NOT_OK_RESULTS should be required to make sense, and
it wont break compatibility, keeping open/close principle in the implementation.

https://progress.opensuse.org/issues/156547