Switch overly verbose "Updating seen from worker" messages to trace
As observed on openqa.suse.de the component openqa-websockets is by far
the systemd unit logging the most. And within openqa-websockets logs the
message "Updating seen from worker" is logged the most. In this commit I
propose to move the message to the "trace" log level so that by default
it would not be logged anymore for instances using a log level of up to
"debug" only. However this means that for idle workers we probably will
not see a message during normal operation anymore only in cases of a
worker picking up jobs and such like this:

```
openqa-websockets-daemon[9246]: [debug] [pid:9246] Started to send message to 2565 for job(s) 17768162
openqa-websockets-daemon[9246]: [debug] [pid:9246] Worker 2480 accepted job 17768169
```

The alternative is to change our openQA instance openqa.suse.de to only
log up to level "info", not "debug".

Related progress issue: https://progress.opensuse.org/issues/182318