Re-add git_auto_commit and improved docs for the git settings
This re-introduces the git_auto_commit setting from #6414. But
with this version, we don't change the default state of the
other settings, or make the 'scm' global setting a master
switch for all of them.

Instead, we maintain compatibility with existing configurations by
allowing 'scm = git' in the global section to enable auto-commit,
if git_auto_commit is not explicitly set to 'no'. Its default
value is ''. So auto-commit defaults to off, but setting *either*
'scm = git' *or* 'git_auto_commit = yes' will turn it on.