Disable git_auto_update by default
As discussed in #6414, #6421, #6426, #6430 and #6472. This isn't
really a safe operation. For instance, in Fedora, the distri is
a git repo, but we commonly interact with it manually as root,
meaning openQA cannot update it (it doesn't have sufficient
permissions). When this fails, it causes problems (guards are
left around which break needle saving in the editor, for e.g.)
It's not even required for the distri dir to be a git repo at
all, but this code will try and update it like one even if it
isn't.

Even aside from the problems, it's safer and more polite to have
behavior like this be disabled by default and enable it on
instances where it's desired.

Signed-off-by: Adam Williamson <awilliam@redhat.com>