Extend single-instance container documentation
https://progress.opensuse.org/issues/183617

This change extends the documentation on the single-instance container.

Choice was made to keep the examples, as they are essential for
understanding how to schedule or clone tests in the container. The
documentation was extended with an example of starting the container and
directly cloning a job, which is the approach displayed by the video in
this section. Also the example job IDs were changed to 1 to avoid
mentioning specific tests which will be unavailable in the future.