Avoid creating stale needle entries in needle editor when symlinks used
When the path of a needle directory contains a symlink we generally still
always refer to the needle under its real path when updating the needles
database table.

The only exception (I am aware of so far) is the needle editor which will
create needle entries in the database for symlinked locations. Those will
never be updated (e.g. `last_seen_time` will never be set) as the code
updating needles always updates the entries for the needles under the
real path. This change removes this exception and thus avoids creating
entries which are never updated.

This is related to https://progress.opensuse.org/issues/184765 as those
entries disturb the sort order.