create_admin: improve search for admin users
The current way of doing this search produces a deprecation
warning:

DBIx::Class::Storage::DBI::select_single(): Query returned more than one row.  SQL that returns multiple rows is DEPRECATED for ->find and ->single at /var/lib/openqa/script/create_admin line 67

Even aside from that, it's a bit weird - why are we relying on
the result of trying to find something that doesn't exist being
equal to 0? That behaviour of `find` isn't documented, AFAICS. It
could just as well return undef or something.

This seems more robust: it's using the appropriate method for
a search which could return 0, 1 or more records, it's explicitly
checking the number of returned records (by putting the search
result in array context and checking the length of the array),
and it doesn't cause any deprecation warnings.

Signed-off-by: Adam Williamson <awilliam@redhat.com>