Avoid showing AJAX	error on index page when user is navigating elsewhere
I tested this locally in Firefox (where the problem is reproducible in
contrast do Chromium) with a `sleep` in `dashboard_build_results` to ensure
reproducibility. With this change the error is no longer shown when the
user is just navigating elsewhere, e.g. clicking on "All tests".

The error is still shown as expected when the web UI is stopped whiel the
AJAX request is still ongoing.

Related ticket: https://progress.opensuse.org/issues/185116