Avoid `Transaction already destroyed` being logged as error
This exception that is normally logged as error is not really due to an
error condition but because the client has already closed the connection so
and thus processing the request is aborted.

This change replaces the default error handler of Mojolicious to avoid this
exception being logged as error and turns it instead into a debug message
which is hopefully easier to understand.

Related ticket: https://progress.opensuse.org/issues/184486