Avoid trying to insert job modules without all required fields
Not sure how to handle the error in the best way. We cannot easily refer
to such an invalid module by its name because the name is whats missing
here in practice. I guess a message like "Skipped … invalid modules" like
I did on the worker side would be possible but since it would be an
additional non-fatal message it would probably not be displayed anywhere
anyway. So maybe it is best to just save the effort of handling this and
rely on the debug messages on the worker-side.

Related ticket: https://progress.opensuse.org/issues/185497