Add function to compute statistics about a scheduled product
This is to be able to add hook script support for scheduled products
because for this we need to be able to determine whether all jobs of a
scheduled product are done (to execute the hook script in this case).

The hook script is supposed to do automatic approval/disapproval of
changes. Hence it needs not only to know whether all jobs are done but also
the state/result they ended up with. So this function returns these kinds
of statistics instead of a binary "all done". Maybe it will need to be
changed to return more high level statistics (like "all passed"), though.

Related ticket: https://progress.opensuse.org/issues/184690