Allow querying the state of scheduled products by distri/version/flavor
This will allow us to check whether all jobs we scheduled for a certain
purpose (e.g. product increment) are done and whether they have passed.

The details of the lookup are explained in the added API documentation.

The query for this is not super efficient as we don't have an index on the
required columns on the scheduled products table. However, it seemed to be
fast enough when I tested this with production data.

Related ticket: https://progress.opensuse.org/issues/184690