Retrieve oldest/newest build results job in the loop
Issue: https://progress.opensuse.org/issues/185494

This just adds back the logic for calculating the oldest time by
checking in the loop which is the oldest, and adds the same for the newest.

Actually it makes it even simpler by just setting the time to the first or last
in the loop, because jobs are sorted by id, which is functionally equal
to sorting by creation time.