Ensure default config of limits/durations make sense
It is possible to configure only one of the limits/durations explicitly,
e.g. `log_storage_duration = 400`. So far the entire results would still be
deleted after 365 days as this is our default for `result_storage_duration`.
So the configuration has no effect. This is probably not what an admin
configuring this had in mind.

With the introduction of `job_storage_duration` this problem is more
apparent as one might have put e.g. `result_storage_duration = 400` or
even `result_storage_duration = 0` into their config file but now we would
still delete the whole job after 365 days as this is the default limit for
`job_storage_duration`.

This change extends the default durations so these settings are still
effective. So if someone put `result_storage_duration = 400` or
`result_storage_duration = 0` in their config file the results would be
kept for 400 days or forever - just as before the introduction of
`job_storage_duration`.

Note that we introduced a validation for this kind of ordering between the
retentions in 026d72a546ac00a600910074b79732ea0809fd83. This change did not
cover the configuration. We should perhaps print a warning in case the
values are invalid. This could be done in the function
`_set_default_storage_durations` introduced by this commit if that is
wanted.

Related ticket: https://progress.opensuse.org/issues/179263